@@ -3,5 +3,5 @@ import { Redirect } from '@docusaurus/router'
 import useBaseUrl from '@docusaurus/useBaseUrl'
 
 export default function Home(): JSX.Element {
-  return <Redirect to={useBaseUrl('/docs/app-router')} />
+  return <Redirect to={useBaseUrl('/docs')} />
 }
